requirements.txt
blinker==1.7.0
Brotli==1.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
Flask==3.0.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
gevent==24.2.1
geventhttpclient==2.2.0
greenlet==3.0.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
locust==2.24.1
locust-plugins==4.4.2
MarkupSafe==2.1.5
msgpack==1.0.8
psutil==5.9.8
pycparser==2.22
pywin32==306
pyzmq==25.1.2
requests==2.31.0
roundrobin==0.0.4
typing_extensions==4.11.0
urllib3==2.2.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
zope.event==5.0
zope.interface==6.3
